+++ services/backend/src/main/java/com/soprasteria/g4it/backend/apiindicator/model/EquipmentImpactBO.java
@@ -34,4 +34,5 @@ public class EquipmentImpactBO {
 
     private Double quantity;
 
+    private Long countValue;
 }