+++ services/frontend/src/app/layout/inventories-footprint/inventories-footprint.component.ts
@@ -124,7 +124,7 @@ export class InventoriesFootprintComponent implements OnInit {
             this.allUnmodifiedFilters[field] = [
                 Constants.ALL,
                 ...uniqueFilterSet[field]
-                    .map((item: any) => (item !== "" ? item : Constants.EMPTY))
+                    .map((item: any) => (item ? item : Constants.EMPTY))
                     .sort(),
             ];
         });